ADD Dialogue system, full mission avaliable
NPC has a functional dialogue UI with different options.
Once the player is in the area and press the suggested button (E), the dialogue and options will appear. Choosing one of the options will show a different text and will setup or not the start of the mission.
Also the text and bool for when the mission is finished are included.