ADD Puzzle
Puzzle is based on 3 game objects with tags that must be placed over 3 plataforms.
This 3 plataforms will detect the correct tag and if so activate a bool.
When the 3 bools are active at the same time the puzzle will be completed and a new bool (doorOpen) of the script SC_Puzzle will give acces to the next area (when added)